Extract nested tests for findMappedName into top level class
TestNG isn't good at running a subset of tests without a log of ceremony
attached (e.g. config files and annotations).
